improve the sidebar layout on small screens
