+++ website/src/icons/FileCodeIcon.vue
@@ -1,0 +1,17 @@
+<template>
+  <!-- from: https://github.com/microsoft/vscode-icons/blob/main/icons/light/file-code.svg -->
+  <svg
+    width="16"
+    height="16"
+    viewBox="0 0 16 16"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      fill-rule="evenodd"
+      clip-rule="evenodd"
+      d="M10.5702 1.14L13.8502 4.44L14.0002 4.8V14.5L13.5002 15H2.50024L2.00024 14.5V1.5L2.50024 1H10.2202L10.5702 1.14ZM10.0002 5H13.0002L10.0002 2V5ZM3.00024 2V14H13.0002V6H9.50024L9.00024 5.5V2H3.00024ZM5.06224 9.533L6.87924 7.705L6.17024 7L4.00024 9.179V9.886L6.17124 12.06L6.87824 11.353L5.06224 9.533ZM8.80024 7.714L9.50024 7.005L11.6892 9.18V9.889L9.50024 12.062L8.79524 11.353L10.6262 9.533L8.80024 7.714Z"
+      fill="var(--svg-icon-fill, black)"
+    />
+  </svg>
+</template>
